refactor: Voucher와 Customer 메뉴 분리 및 실행 책임 분담
